+++ CHANGELOG.md
@@ -370,6 +370,8 @@
 * Updated to Unity 2020.3 LTS
 * Updated package versions
 * Fixed UnityWebRequestAssetImporter using obsolete error checks
+* ActionSpecialSplitter now catches exceptions in called action specials by default
+* ActorController now catches exceptions thrown by OnDeathSpecial
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]